+++ backend/rentAccess/jwtauth/urls.py
@@ -1,4 +1,4 @@
-from .views import registration, login
+from .views import registration, login, logout, logout_all
 from django.urls import path
 from rest_framework_simplejwt import views as jwt_views
 from rest_framework_simplejwt.views import TokenVerifyView
@@ -8,4 +8,6 @@ urlpatterns = [
     path('token/', login, name='token_obtain_pair'),
     path('token/refresh/', jwt_views.TokenRefreshView.as_view(), name='token_refresh'),
     path('token/verify/', TokenVerifyView.as_view(), name='token_verify'),
+    path('logout/', logout, name='logout'),
+    path('logoutall/', logout_all, name='logout_all'),
 ]