+++ backend/rentAccess/jwtauth/views.py
@@ -1,14 +1,14 @@
 from django.contrib.auth import get_user_model
 from rest_framework import response, decorators, permissions, status
-from rest_framework.permissions import AllowAny
-from rest_framework_simplejwt.tokens import RefreshToken
-from rest_framework_simplejwt.views import TokenObtainPairView
+# from rest_framework.permissions import AllowAny
+from rest_framework.response import Response
+from rest_framework_simplejwt.token_blacklist.models import OutstandingToken, BlacklistedToken
+from rest_framework_simplejwt.tokens import RefreshToken, AccessToken
 
 from .serializers import UserCreateSerializer, UserLoginSerializer
 from userAccount.models import Profile
 
-# TODO: create new accounts when user registers
-# TODO: add user and role to response
+# TODO: create blacklist for tokens
 
 
 User = get_user_model()
@@ -18,7 +18,7 @@ User = get_user_model()
 @decorators.permission_classes([permissions.AllowAny])
 def registration(request):
 	serializer = UserCreateSerializer(data=request.data)
-	permission_classes = (AllowAny,)
+	# permission_classes = (AllowAny,)
 	if not serializer.is_valid():
 		return response.Response(serializer.errors, status.HTTP_400_BAD_REQUEST)
 	user = serializer.save()
@@ -43,7 +43,7 @@ def registration(request):
 @decorators.permission_classes([permissions.AllowAny])
 def login(request):
 	serializer = UserLoginSerializer(data=request.data)
-	permission_classes = (AllowAny,)
+	# permission_classes = (AllowAny,)
 	if not serializer.is_valid():
 		return response.Response(serializer.errors, status.HTTP_400_BAD_REQUEST)
 	user = serializer.save()
@@ -62,4 +62,28 @@ def login(request):
 			"account_type": str(profile.account_type),
 		},
 	}
-	return response.Response(res, status.HTTP_201_CREATED)
+	return response.Response(res, status.HTTP_200_OK)
+
+
+@decorators.api_view(["POST"])
+@decorators.permission_classes([permissions.IsAuthenticated])
+def logout(request):
+	try:
+		refresh_token = request.data["refresh"]
+		token = RefreshToken(refresh_token)
+		token.blacklist()
+		return Response(status=status.HTTP_205_RESET_CONTENT)
+	except Exception as e:
+		return Response(status=status.HTTP_400_BAD_REQUEST)
+
+
+@decorators.api_view(["POST"])
+@decorators.permission_classes([permissions.IsAuthenticated])
+def logout_all(request):
+	tokens = OutstandingToken.objects.filter(user_id=request.user.id)
+	for token in tokens:
+		t, _ = BlacklistedToken.objects.get_or_create(token=token)
+
+	return Response(status=status.HTTP_205_RESET_CONTENT)
+
+